backend/requirements.txt
# FastAPI and Server
fastapi==0.121.2
uvicorn[standard]==0.32.0
pydantic==2.9.2
starlette==0.49.3

# Google Generative AI
google-genai==0.2.2

# Environment Variables
python-dotenv==1.0.1

# HTTP Requests
httpx==0.27.2
aiohttp==3.10.10
requests==2.32.3

# GitHub Ingestion
gitingest==0.3.1

# Other utilities
typing-extensions==4.12.2
tiktoken==0.12.0
